// shippingSystemUi/ShippingUI/src/app/Core/Models/Permission.ts
export enum Government {
  Read = 'permission.goverments.read',
  Create = 'permission.goverments.add',
  Update = 'permission.goverments.update',
  Delete = 'permission.goverments.delete',
}

export enum City {
  Read = 'permission.cities.read',
  Create = 'permission.cities.add',
  Update = 'permission.cities.update',
  Delete = 'permission.cities.delete',
}

export enum Trader {
  Read = 'permission.traders.read',
  Create = 'permission.traders.add',
  Update = 'permission.traders.update',
  Delete = 'permission.traders.delete',
}

export enum Employee {
  Read = 'permission.employees.read',
  Create = 'permission.employees.add',
  Update = 'permission.employees.update',
  Delete = 'permission.employees.delete',
}

export enum Branch {
  Read = 'permission.branches.read',
  Create = 'permission.branches.add',
  Update = 'permission.branches.update',
  Delete = 'permission.branches.delete',
}

export enum Sales {
  Read = 'permission.sales.read',
  Create = 'permission.sales.add',
  Update = 'permission.sales.update',
  Delete = 'permission.sales.delete',
}

export enum Order {
  Read = 'permission.orders.read',
  Create = 'permission.orders.add',
  Update = 'permission.orders.update',
  Delete = 'permission.orders.delete',
}

export enum Role {
  Read = 'permission.permissions.read',
  Create = 'permission.permissions.add',
  Update = 'permission.permissions.update',
  Delete = 'permission.permissions.delete',
}